Re-think makefile approach
After using this in a few personal projects I've realized that this is
better served as an include to projects which have an existing Makefile
and template setup. At the same time, I still wanted this to be a
stand-alone solution.

Readme updated with better instructions to demonstrate the above.